feat: add `working-directory` as an input (#359)
* feat: add `working-directory` as an input

Support customizing the `working-directory` that promptfoo is run using.

For example, you can set `promptfoo-version` to an empty string to use
any valid version that is already installed and available in the
`working-directory`, e.g. if you had run `npm install` in a previous
step.

```yaml
      - name: Run promptfoo evaluation
        uses: promptfoo/promptfoo-action@v1
        with:
          promptfoo-version: ''
          working-directory: './my-folder'
```

* squash! feat: add `working-directory` as an input

Run `npm run package`.