docs: Clarify version requirements for workflow_dispatch and push events
Fixes #703

The v1 release does not support workflow_dispatch or push events.
These features were added after the v1.0.0 release and are only
available when using @main.

Added a clear note in the README to inform users that they need
to use @main instead of @v1 to access these features.